requirements.txt
requests>=2.28.0
Pillow>=10.0.0
zxing-cpp>=2.0.0
